[update] renamed cfp-tag.ts
